+++ config/config.go
@@ -27,19 +27,19 @@ func (o OSEnvGetter) Getbool(key string) (bool, error) {
 
 const (
 	ghTokenKey       = "STARFEED_GITHUB_API_TOKEN"
-	freshRssUrlKey   = "STARFEED_FRESHRSS_URL"
-	freshRssUserKey  = "STARFEED_FRESHRSS_USER"
-	freshRssTokenKey = "STARFEED_FRESHRSS_API_TOKEN"
+	freshRSSURLKey   = "STARFEED_FRESHRSS_URL"
+	freshRSSUserKey  = "STARFEED_FRESHRSS_USER"
+	freshRSSTokenKey = "STARFEED_FRESHRSS_API_TOKEN"
 	debugModeKey     = "STARFEED_DEBUG_MODE"
 	singleRunModeKey = "STARFEED_SINGLE_RUN_MODE"
 	httpTimeoutKey   = "STARFEED_HTTP_TIMEOUT"
 )
 
 type Config struct {
-	GithubToken   string
-	FreshRssUrl   string
-	FreshRssUser  string
-	FreshRssToken string
+	GitHubToken   string
+	FreshRSSURL   string
+	FreshRSSUser  string
+	FreshRSSToken string
 	DebugMode     bool
 	SingleRunMode bool
 	HttpTimeout   time.Duration
@@ -48,9 +48,9 @@ type Config struct {
 func NewConfig(envGetter EnvGetter) (*Config, error) {
 	// Check for required environment variables
 	if envGetter.Getenv(ghTokenKey) == "" ||
-		envGetter.Getenv(freshRssUrlKey) == "" ||
-		envGetter.Getenv(freshRssUserKey) == "" ||
-		envGetter.Getenv(freshRssTokenKey) == "" {
+		envGetter.Getenv(freshRSSURLKey) == "" ||
+		envGetter.Getenv(freshRSSUserKey) == "" ||
+		envGetter.Getenv(freshRSSTokenKey) == "" {
 		return nil, errors.New("missing required environment variables")
 	}
 
@@ -72,10 +72,10 @@ func NewConfig(envGetter EnvGetter) (*Config, error) {
 	}
 
 	return &Config{
-		GithubToken:   envGetter.Getenv(ghTokenKey),
-		FreshRssUrl:   envGetter.Getenv(freshRssUrlKey),
-		FreshRssUser:  envGetter.Getenv(freshRssUserKey),
-		FreshRssToken: envGetter.Getenv(freshRssTokenKey),
+		GitHubToken:   envGetter.Getenv(ghTokenKey),
+		FreshRSSURL:   envGetter.Getenv(freshRSSURLKey),
+		FreshRSSUser:  envGetter.Getenv(freshRSSUserKey),
+		FreshRSSToken: envGetter.Getenv(freshRSSTokenKey),
 		DebugMode:     debugMode,
 		SingleRunMode: singleRunMode,
 		HttpTimeout:   httpTimeout,