+++ github/github.go
@@ -15,7 +15,7 @@ import (
 // for all starred repos.
 type GitHubStarredFeedBuilder interface {
 	GetStarredRepos() (map[string]GitHubRepo, error)
-	IsGithubReleasesFeed(feedUrl string) bool
+	IsGitHubReleasesFeed(feedUrl string) bool
 }
 
 // gitHubStarredFeedBuilder is the private implementation of GitHubStarredFeedBuilder.
@@ -49,7 +49,7 @@ func NewGitHubStarredFeedBuilder(
 func (gh *gitHubStarredFeedBuilder) GetStarredRepos() (map[string]GitHubRepo, error) {
 	allFeeds := make(map[string]GitHubRepo)
 	getUrl := "https://api.github.com/user/starred?per_page=100"
-	slog.Debug("Querying Github for starred repos", "url", getUrl)
+	slog.Debug("Querying GitHub for starred repos", "url", getUrl)
 
 	for {
 		ghResponse, err := gh.doApiRequest(getUrl)
@@ -77,17 +77,17 @@ func (gh *gitHubStarredFeedBuilder) GetStarredRepos() (map[string]GitHubRepo, er
 	}
 }
 
-// This function returns true if a repoUrl is a Github release repo
+// This function returns true if a repoUrl is a GitHub release repo
 // Arguments:
 // - feedUrl: The URL of the RSS feed to check.
-func (gh *gitHubStarredFeedBuilder) IsGithubReleasesFeed(feedUrl string) bool {
+func (gh *gitHubStarredFeedBuilder) IsGitHubReleasesFeed(feedUrl string) bool {
 	return gh.isRelRepoRegex.MatchString(feedUrl)
 }
 
-func (gh *gitHubStarredFeedBuilder) doApiRequest(url string) (*GithubResponse, error) {
+func (gh *gitHubStarredFeedBuilder) doApiRequest(url string) (*GitHubResponse, error) {
 	headers := map[string]string{
 		"Authorization":        fmt.Sprintf("Bearer %s", gh.token),
-		"X-Github-Api-Version": "2022-11-28",
+		"X-GitHub-Api-Version": "2022-11-28",
 		"User-Agent":           "github.com/atomicmeganerd/starfeed",
 		"Content-Type":         "application/json",
 		"Accept":               "application/json",
@@ -113,7 +113,7 @@ func (gh *gitHubStarredFeedBuilder) doApiRequest(url string) (*GithubResponse, e
 		return nil, fmt.Errorf("github returned an http error code %d", res.StatusCode)
 	}
 
-	ghResponse, err := gh.processGithubResponse(res)
+	ghResponse, err := gh.processGitHubResponse(res)
 	if err != nil {
 		return nil, err
 	}
@@ -121,9 +121,9 @@ func (gh *gitHubStarredFeedBuilder) doApiRequest(url string) (*GithubResponse, e
 	return ghResponse, nil
 }
 
-func (gh *gitHubStarredFeedBuilder) processGithubResponse(
+func (gh *gitHubStarredFeedBuilder) processGitHubResponse(
 	r *http.Response,
-) (*GithubResponse, error) {
+) (*GitHubResponse, error) {
 	data, err := io.ReadAll(r.Body)
 	if err != nil {
 		return nil, err
@@ -134,9 +134,9 @@ func (gh *gitHubStarredFeedBuilder) processGithubResponse(
 	for link := range links {
 		matches := gh.nextPageLinkRegex.FindStringSubmatch(link)
 		if len(matches) == 2 {
-			return &GithubResponse{data: data, nextPage: matches[1]}, nil
+			return &GitHubResponse{data: data, nextPage: matches[1]}, nil
 		}
 	}
 
-	return &GithubResponse{data: data}, nil
+	return &GitHubResponse{data: data}, nil
 }