+++ CHANGELOG.md
@@ -38,7 +38,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
-- Fixed a bug that causes Starfeed to unsubscribe from all non-Github release feeds.
+- Fixed a bug that causes Starfeed to unsubscribe from all non-GitHub release feeds.
 - Added some initial test coverage.
 
 ## [0.1.1] - 2024-12-24
@@ -91,7 +91,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Publishing the Atom feeds from Github to FreshRSS works!
+- Publishing the Atom feeds from GitHub to FreshRSS works!
 
 ## [0.0.1] - 2024-12-12